Azure Functions/GPT4V_Custom_Skill/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions == 1.19.0
openai == 1.25.2